Implement Zeta placeOrderV3. Add ZetaTest.

+++ zeta/src/main/java/com/mmorrell/zeta/model/ZetaOrderType.java
@@ -10,8 +10,8 @@ public enum ZetaOrderType {
         this.value = value;
     }
 
-    public int getValue() {
-        return value;
+    public byte getValue() {
+        return (byte) value;
     }
 
 }
