+++ zeta/src/main/java/com/mmorrell/zeta/model/ZetaSide.java
@@ -10,7 +10,7 @@ public enum ZetaSide {
         this.value = value;
     }
 
-    public int getValue() {
-        return value;
+    public byte getValue() {
+        return (byte) value;
     }
 }
